moving nested classes to separate files in TxPool
